Wrote serial_read/write functions removed IRQ based echo ring buffer
The serial_read/write functions are not very efficient for single
characters because it turnes the interrupts on and off for every
character. It might be better to refactor this using a ring buffer,
and have serial read/write use the ring instead of passing the buffer
directly, and thrashing the IRQ. A ring buffer provides better latency,
but the current implementation, is zero-copy with better through put.

This commit has the bufecho ring buffer implementation:
  6349c2d8203355a7ca1523572d9e58e02f54eb05