Removed PID error scaling based on range
Previously we scaled the PID error value to as a value from 0-1 over
the entire calibration range.  This causes PID scaling problems when
the range changes, so now we report error to the PID controller as
the distance between the target and current position.

This will make the PID tunables (kp, ki, kd) much smaller, but it
will solve the calibration scale problem.