Added rotor speed exponents
Internally, speeds are values from 0-1. Thus, if we raise the speed to a power,
then smaller values are biased toward even smaller values. The motor speed
returned by the pid controller is raised to this exponent before setting the
motor speed. This allows the motor to slow down as it approaches it's target,
and minimizes near-target jitter. The default value of 1.0 is linear, thus
using the pid controller without change. We find a value between 1.1 and 2.0 to
be effective.