Fixed FF pid control
The Feed-forward part of the pid controller (Kvff/Kaff) responds to differences
in the target (control input), like when tracking a satellite. Since the
satellite update code is slower than the pid controller (20/sec vs 100/sec),
the difference between the previous target command is zero for 80% of the time.
This change iterates to look back to the most recent target that is different
from the current target to keep FF smooth.