Make ads111x_t cast compatible with i2c_req_t
We were having to recrate the ads111x_t structure from the channel and vref
information in the rotor structure because the ads111x_t config register data
was not available when all we had was a i2c_req_t object. This refactor appends
the ads111x config to the i2c_req_t struct so it is available simply by casting
to ads111x_t* . It would be a good idea to do this for future i2c drivers as
well.

Also, we removed the syncronous measure funcionality since the ADCs are always
measured from the ADC interrupt.