Continue motor for 3 seconds on error
If a motor is in a NAN state, then increment error_count (100 ticks/sec).
While the error count is less than 300 (3 seconds), keep the motor moving the
same speed.

This is a work-around for bad spots in potentiometers.  The motor just needs to
get past it.  Once it does, the position tracking will take over and correct.
If the error persists beyond 300 ticks then it will stop the motor.